% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beads_bam_bw.R
\name{beads_bam_bw}
\alias{beads_bam_bw}
\title{BEADS internal}
\usage{
beads_bam_bw(bam.file, bw.control, bw.mappability, genome, uniq = TRUE,
  insert = 200L, mapq_cutoff = 10L, export = "BEADS", rdata = FALSE,
  export_er = TRUE, quickMap = TRUE)
}
\description{
The function runs BEADS for signature BAM, BAM, BW, ANY: \cr
\code{signature(experiment='BamFile', control='BamFile', mappability='BigWigFile', genome='ANY')}
}
\details{
This function is package internal and should not be executed directly
by users.
}
\keyword{internal}

