% Generated by roxygen2 (4.0.0): do not edit by hand
\name{beads_batch}
\alias{beads_batch}
\title{Main rBEADS wrapper}
\usage{
beads_batch(config = "BeadsConfig.csv")
}
\arguments{
\item{config}{Configuration file in CSV format.}
}
\value{
LogFile - Read the log file for further run info.
}
\description{
This function should be started ni the BAM/Solexa data directory.
It inputs CSV config file and run BEADS algoritm with specified settings.
}
\details{
rBEADS configuration file are 4 columns Comma Separated Values (CSV) tables. They can be created using any spreadsheet editor or plain text editor. The structure of sample input file:
 \preformatted{
   Sample,Control,rep,ER
   ChIPFile1.bam,InputFile1.bam,rep1,ERfile.bed
   ChIPFile2.bam,FRM,rep2a,none
   ChIPFile3.bam,EGS,rep2b,auto
 }

 The tabular structure:
 \tabular{llll}{
   Sample \tab Control \tab rep \tab ER \cr
   ChIPFile1.bam \tab InputFile1.bam \tab rep1 \tab ERfile.bed \cr
   ChIPFile2.bam \tab FRM \tab rep2a \tab none \cr
   ChIPFile3.bam \tab EGS \tab rep2b \tab auto \cr
  }

 The \code{Sample} column contains ChIP experiment aligned short reads files.

 The \code{Control} column contains Input file corresponding to ChIP experiment. This filed may hold 2 special values - \code{EGS} forces BEADS wrapper to use pre-calculate EGS summed input track shipped with R package. Similarly \code{FRM} value indicates using summed formaldehyde input (also available in rBEADS package).

 \code{rep} column is not currently used by and have been incorporated to easier organize coning table. This filed can hold any value (excluding commas).

 The last column - \code{ER} - indicates which enriched regions should be used in BEADS run. The \code{auto} value indicates build in rBEADS algorithm for finding enriched regions, while \code{none} switches off enriched regions masking from the algorithm (useful for testing input files). Finally, if user decides to use external enriched regions finder (e. g. by peak calling) this field may take BED file name.
}
\examples{
source("http://bioconductor.org/biocLite.R")
biocLite("BSgenome.Celegans.UCSC.ce10")

require(BSgenome.Celegans.UCSC.ce10)
require(rbeads)
beads(config='BeadsConfig.csv')
}
\author{
Przemyslaw Stempor
}
\references{
http://beads.sourceforge.net/
}
\keyword{beads}

