% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{rbeads}
\alias{rbeads}
\alias{rbeads-package}
\title{rbeads - R implementation of Bias Elimination Algorithm for Deep Sequencing}
\description{
BEADS is a normalization scheme that corrects nucleotide composition bias, mappability variations and differential local DNA structural effects in deep sequencing data. In high-throughput sequencing data, the recovery of sequenced DNA fragments is not uniform along the genome. In particular, GC-rich sequences are often over-represented and AT-rich sequences under-represented in sequencing data. In addition, the read mapping procedure also generates regional bias. Sequence reads that can be mapped to multiple sites in the genome are usually discarded. Genomic regions with high degeneracy therefore show lower mapped read coverage than unique portions of the genome. Mappability varies along the genome and thus creates systematic bias. Futhermore, local DNA or chromatin structural effects can lead to coverage inhomogeneity of sequencing data.
}
\details{
\tabular{ll}{
 Package: \tab rBEADS\cr
 Type: \tab Package\cr
 Version: \tab 1.0\cr
 Date: \tab 2011-07-04\cr
 License: \tab LGPL\cr
 LazyLoad: \tab yes\cr
}
}
\examples{
example(beads)
}
\author{
Przemyslaw Stempor
}
\references{
http://beads.sourceforge.net/manual.php
}
\keyword{CHiP-seq}
\keyword{beads}

